safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEDisplacementMapElement-svgdom-yChannelSelector-prop-expected.txt

CONSOLE MESSAGE: line 1: ReferenceError: Can't find variable: SVGFEDisplacementMapElement
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'yChannelSelector' property of the SVGFEDisplacementMapElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
